+++ python_training/Midas/midas/__init__.py
@@ -1,4 +1,4 @@
+from midas.core.engine import engine
 from midas.main import insert_info
 from midas.queries.queries import *
-from midas.core.utils import get_session
-from midas.core.engine import engine
+from midas.utils.utils import get_session
